requirements.txt
streamlit>=1.31.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-docx>=1.1.0
PyPDF2>=3.0.1
requests>=2.31.0